tests: semaphore: categorize the semaphore test cases
Categorize the semaphore test cases into three group:

The first group is the module testing, target to verify the
semaphore API itself. Ex. testsuite: semaphore_api.

The second group is the integration testing, target to verify
the semaphore works with other modules or usage in the system.
Ex. testsuite: sys_sem.

The third group is fault injection testing, target to passing
incorrect parameters to semaphore, to verify the errors
can be detected and be handled.
Ex. testsuite: sem_error_case
