+++ tests/kernel/semaphore/sem_error_case/CMakeLists.txt
@@ -4,9 +4,5 @@ cmake_minimum_required(VERSION 3.20.0)
 find_package(Zephyr REQUIRED HINTS $ENV{ZEPHYR_BASE})
 project(semaphore)
 
-target_include_directories(app PRIVATE
-  ../common_files/include/
-  )
-
 FILE(GLOB app_sources src/*.c)
 target_sources(app PRIVATE ${app_sources})
