+++ tests/kernel/semaphore/sem_error_case/prj.conf
@@ -1,3 +1,3 @@
 CONFIG_ZTEST=y
-CONFIG_IRQ_OFFLOAD=y
 CONFIG_TEST_USERSPACE=y
+CONFIG_ZTEST_FATAL_HOOK=y
