+++ tests/kernel/semaphore/sem_error_case/testcase.yaml
@@ -1,0 +1,4 @@
+tests:
+  kernel.semaphore.error:
+    filter: CONFIG_ARCH_HAS_USERSPACE
+    tags: kernel userspace ignore_faults
