+++ tests/kernel/semaphore/semaphore/prj.conf
@@ -1,4 +1,2 @@
 CONFIG_ZTEST=y
-CONFIG_IRQ_OFFLOAD=y
 CONFIG_TEST_USERSPACE=y
-CONFIG_ZTEST_FATAL_HOOK=y
