+++ tests/kernel/semaphore/semaphore/testcase.yaml
@@ -1,3 +1,3 @@
 tests:
-  kernel.semaphore:
-    tags: kernel userspace ignore_faults
+  kernel.semaphore.api:
+    tags: kernel userspace
