+++ tests/kernel/semaphore/sys_sem/testcase.yaml
@@ -1,4 +1,4 @@
 tests:
   kernel.semaphore.usage:
     filter: CONFIG_ARCH_HAS_USERSPACE
-    tags: kernel userspace ignore_faults
+    tags: kernel userspace
